requirements.txt
alabaster==0.7.10
anaconda-client==1.6.3
anaconda-navigator==1.6.4
anaconda-project==0.6.0
apiai==1.2.3
appnope==0.1.0
appscript==1.0.1
APScheduler==3.3.1
asn1crypto==0.22.0
astroid==1.4.9
astropy==1.3.2
attrs==17.2.0
Automat==0.6.0
autopep8==1.3.4
Babel==2.4.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
blaze==0.10.1
bleach==1.5.0
bokeh==0.12.5
boto==2.48.0
Bottleneck==1.2.1
bs4==0.0.1
bz2file==0.98
cachetools==2.0.1
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cloudpickle==0.2.2
clyent==1.2.2
colorama==0.3.9
conda==4.4.8
constantly==15.1.0
contextlib2==0.5.5
cryptography==1.8.1
cssselect==1.0.1
cycler==0.10.0
Cython==0.25.2
cytoolz==0.8.2
dask==0.14.3
datashape==0.5.4
DAWG==0.7.8
DAWG-Python==0.7.2
decorator==4.1.2
definitions==0.2.0
dialogflow==0.1.0
distributed==1.16.3
docopt==0.6.2
docutils==0.13.1
dominate==2.3.1
echarts-python==0.1.3
entrypoints==0.2.2
esl==0.5.1
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==0.12.2
Flask-Cors==3.0.2
Flask-MySQL==1.4.0
Flask-SocketIO==2.9.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
FreeSWITCH-ESL-Python==1.2
future==0.16.0
fuzzywuzzy==0.15.1
gensim==2.3.0
gevent==1.2.1
google-api-core==0.1.4
google-auth==1.3.0
googleapis-common-protos==1.5.3
greenlet==0.4.12
grpcio==1.8.4
gTTS==1.2.2
gTTS-token==1.1.1
gunicorn==19.7.1
h5py==2.7.0
HeapDict==1.0.0
html5lib==0.999
hyperlink==17.3.1
idna==2.6
imagesize==0.7.1
importer==1.2
incremental==17.5.0
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.2.5
itsdangerous==0.24
JACK-Client==0.4.3
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
keen==0.4.0
lazy-object-proxy==1.2.2
line-profiler==2.0
llvmlite==0.18.0
locket==0.2.0
lxml==3.7.3
MarkupSafe==0.23
matplotlib==2.0.2
mistune==0.7.4
-e git+https://gitlab.com/chamberijorik/mvt.git@7ebd1317e0cbe46f2bbb46ebb383ae2a9c8018d9#egg=movieterra
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.9
navigator-updater==0.1.0
nbconvert==5.1.1
nbformat==4.3.0
networkx==1.11
nexmo==2.0.0
nltk==3.2.4
nose==1.3.7
notebook==5.0.0
numba==0.33.0
numexpr==2.6.2
numpy==1.13.1
numpydoc==0.6.0
odo==0.5.0
olefile==0.44
opencv-python==3.3.0.10
openpyxl==2.4.7
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.1
parsel==1.2.0
partd==0.3.8
pathlib2==2.2.1
patsy==0.4.1
pcapy==0.11.1
peewee==2.10.2
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
pipreqs==0.4.9
playsound==1.2.2
plotly==2.0.12
ply==3.10
prompt-toolkit==1.0.14
protobuf==3.5.1
psutil==5.2.2
ptyprocess==0.5.1
py==1.4.33
pyasn1==0.3.4
pyasn1-modules==0.2.1
PyAudio==0.2.11
pyclust==0.2.0
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.17
pycrypto==2.6.1
pycryptodome==3.4.7
pycurl==7.43.0
PyDispatcher==2.0.5
pydub==0.20.0
pyflakes==1.5.0
pygame==1.9.3
Pygments==2.2.0
PyJWT==1.5.3
pylint==1.6.4
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
PyMySQL==0.7.11
pyodbc==4.0.16
pyOpenSSL==17.0.0
pyparsing==2.1.4
pypcap==1.2.0
pytest==3.0.7
python-dateutil==2.6.0
python-engineio==1.7.0
python-socketio==1.8.3
python-vlc==3.0.101
pyttsx3==2.7
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.0
QtPy==1.2.1
queuelib==1.4.2
requests==2.18.4
rope-py3k==0.9.4.post1
rsa==3.4.2
scikit-fuzzy==0.3
scikit-image==0.13.0
scikit-learn==0.18.1
scipy==0.19.1
Scrapy==1.4.0
seaborn==0.7.1
service-identity==17.0.0
sets==0.3.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
smart-open==1.5.3
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
sounddevice==0.3.10
Sphinx==1.5.6
spyder==3.2.3
SQLAlchemy==1.1.9
statsmodels==0.8.0
sympy==1.0
tables==3.3.0
tblib==1.3.2
terminado==0.6
testpath==0.3
tmdbsimple==1.7.0
toolz==0.8.2
tornado==4.5.1
tqdm==4.15.0
traitlets==4.3.2
translate==2.0.0
translation==1.0.5
transliterate==1.10
treelib==1.4.0
tsfresh==0.10.1
tts==0.1
Twisted==17.5.0
tzlocal==1.4
unicodecsv==0.14.1
urllib3==1.22
virtualenv==15.1.0
vk==2.0.2
w3lib==1.18.0
Wand==0.4.4
Wave==0.0.2
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==2.0.0
wit==4.3.0
wordcloud==1.3.1
wrapt==1.10.10
WTForms==2.1
xlrd==1.0.0
XlsxWriter==0.9.6
xlwings==0.10.4
xlwt==1.2.0
xvfbwrapper==0.2.9
yarg==0.1.9
zict==0.1.2
zope.interface==4.4.2
